Mock external comm
Refactoring project structure (including AbstractScanning to __init__.py), config for external comm, getting rid of stytra terms straggling here and there, black formatting
