+++ sashimi/hardware/external_trigger/__init__.py
@@ -1,0 +1,6 @@
+from typing import Optional
+
+
+class AbstractComm:
+    def trigger_and_receive_duration(self, config) -> Optional[float]:
+        return None
