+++ sashimi/processes/external_communication.py
@@ -1,49 +1,13 @@
 from multiprocessing import Queue
 from sashimi.processes.logging import LoggingProcess
+from sashimi.utilities import clean_json
 from sashimi.events import LoggedEvent
+from sashimi.config import read_config
+from sashimi.hardware import external_comm_class_dict
 from queue import Empty
-import zmq
-from dataclasses import asdict, is_dataclass
-from enum import Enum
-from typing import Optional
 
 
-def clean_json(d):
-    if isinstance(d, dict):
-        cleaned = dict()
-        for key, value in d.items():
-            cleaned[key] = clean_json(value)
-        return cleaned
-    elif isinstance(d, Enum):
-        return d.name
-    elif is_dataclass(d):
-        return clean_json(asdict(d))
-    else:
-        return d
-
-
-class AbstractComm:
-    def trigger_and_receive_duration(self, config) -> Optional[float]:
-        return None
-
-
-class StytraComm(AbstractComm):
-    def __init__(self, computer_adderss):
-        self.zmq_tcp_address = computer_adderss
-
-    def trigger_and_receive_duration(self, config) -> Optional[float]:
-        zmq_context = zmq.Context()
-        with zmq_context.socket(zmq.REQ) as zmq_socket:
-            zmq_socket.connect(self.zmq_tcp_address)
-            zmq_socket.send_json(config)
-            poller = zmq.Poller()
-            poller.register(zmq_socket, zmq.POLLIN)
-            duration = None
-            if poller.poll(1000):
-                duration = zmq_socket.recv_json()
-
-        zmq_context.destroy()
-        return duration
+conf = read_config()
 
 
 class ExternalComm(LoggingProcess):
@@ -53,17 +17,22 @@ class ExternalComm(LoggingProcess):
         experiment_start_event: LoggedEvent,
         is_saving_event: LoggedEvent,
         is_waiting_event: LoggedEvent,
-        stytra_address="tcp://O1-589:5555",
+        address=conf["external_communication"]["address"],
         scanning_trigger=True,
     ):
-        super().__init__(name="stytracomm")
+        super().__init__(name="external_comm")
         self.current_settings_queue = Queue()
         self.current_settings = None
-        self.start_stytra = experiment_start_event.new_reference(self.logger)
+        self.start_comm = experiment_start_event.new_reference(self.logger)
         self.stop_event = stop_event.new_reference(self.logger)
         self.saving_event = is_saving_event.new_reference(self.logger)
         self.duration_queue = Queue()
-        self.comm = StytraComm(stytra_address)
+        if conf["scopeless"]:
+            self.comm = external_comm_class_dict["test"]()
+        else:
+            self.comm = external_comm_class_dict[
+                conf["external_communication"]["name"]
+            ](address)
         self.scanning_trigger = scanning_trigger
         if self.scanning_trigger:
             self.waiting_event = is_waiting_event.new_reference(self.logger)
@@ -71,7 +40,7 @@ class ExternalComm(LoggingProcess):
     def trigger_condition(self):
         if self.scanning_trigger:
             return (
-                self.start_stytra.is_set()
+                self.start_comm.is_set()
                 and self.saving_event.is_set()
                 and not self.waiting_event.is_set()
             )
@@ -92,5 +61,5 @@ class ExternalComm(LoggingProcess):
                 if duration is not None:
                     self.duration_queue.put(duration)
                 self.logger.log_message("sent communication")
-                self.start_stytra.clear()
+                self.start_comm.clear()
         self.close_log()
