+++ sashimi/processes/scanning.py
@@ -7,7 +7,7 @@ from sashimi.hardware.scanning.scanloops import (
     PlanarScanLoop,
     VolumetricScanLoop,
 )
-from sashimi.hardware.scanning.interface import ScanningError
+from sashimi.hardware.scanning.__init__ import ScanningError
 from sashimi.hardware.scanning.mock import open_mockboard
 
 try:
